handle KeyErrors on removal better
When attempting to delete a child that didn't exist in the parent, the
previous behavior was to silently ignore the KeyError. This has been
changed to throw a warning, since the only way I know that this can
happen is that either a child wasn't attached to a given parent in the
first place, or that child had its name changed after being attached to
a parent.

Both of those cases are, for the moment, unexpected/unsupported.